♻️ refactor:  SearchReposStateをState層に移動
Issue #285
