+++ lib/domain/state/search_repos.freezed.dart
